+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 This log tries to follow the good principles of [Keep a CHANGELOG](http://keepachangelog.com/).
 
 
+## 2.1.1 - 2024-11-05
+
+### Added
+- `odsPicto`: New `local-id` parameter that is used to clone a reference SVG prexisting in the page,
+  to avoid having to fetch the SVG remotely.
+- `odsSocialButtons`: Twitter logo has been replaced with the new X logo.
+- `odsTable`: Every text field is now sortable.
+
 ## 2.1.0 - 2023-10-03
 This version notably adds the support of jQuery 3, while keeping compatibility with jQuery 2 as well.